+++ docs/02-validator-adapters.md
@@ -1,7 +1,7 @@
 # Validator Adapters
 
-Create adapters for existing validation libraries! Included is a basic abstract class to create a new adapter,
-see [`Graze\DataValidator\Adapter\AbstractExecutableAdapter`](../src/Adapter/AbstractExecutableAdapter.php), and an adapter for the `Respect\Validation` library.
+Create adapters for existing validation libraries! Included is a basic abstract class to create a new validator,
+see [`Graze\DataValidator\Validator\AbstractExecutableValidator`](../src/Validator/AbstractExecutableValidator.php), and an adapter for the `Respect\Validation` library.
 
 When writing your own, ensure that the `validate` method returns a suitable string when validation fails.
